Creating Authorization policies

+++ Blog-Website/app/Http/Controllers/PostController.php
@@ -9,7 +9,7 @@ use Illuminate\Http\Request;
 class PostController extends Controller
 {
     public function index(){
-        $posts = Post::all();
+        $posts = auth()->user()->posts()->paginate(5);
         return view('admin.posts.index',['posts'=>$posts]);
     }
     public function show(Post $post){
@@ -19,10 +19,11 @@ class PostController extends Controller
         return view('admin.posts.create');
     }
     public function store(){
-       $inputs = request()->validate([
-        'title'=>'required|min:8|max:255',
-        'post_image'=>'file',
-        'body'=>'required',
+        $this->authorize('create',Post::class);
+        $inputs = request()->validate([
+            'title'=>'required|min:8|max:255',
+            'post_image'=>'file',
+            'body'=>'required',
        ]);
        if(request('post_image')){
         $inputs['post_image'] = request('post_image')->store('images');
@@ -32,9 +33,11 @@ class PostController extends Controller
        return redirect()->route('post.index');
     }
     public function edit(Post $post){
+        $this->authorize('view',$post);
         return view('admin.posts.edit',['post'=>$post]);
     }
     public function destroy(Post $post,Request $request){
+        $this->authorize('delete',$post);
         $post->delete();
         $request->session()->flash('message','Post was deleted');
         return back();
@@ -51,6 +54,7 @@ class PostController extends Controller
            }
            $post->title = $inputs['title'];
            $post->body = $inputs['body'];
+           $this->authorize('update',$post);
            $post->save();
            session()->flash('post-updated-message','Post with title was updated '.$inputs['title']);
            return redirect()->route('post.index');
